refactor(cli-dxns): Refactor DXNS commands
