+++ packages/cli-dxns/src/modules/commands/index.ts
@@ -1,0 +1,11 @@
+//
+// Copyright 2020 DXOS.org
+//
+
+export * from './account';
+export * from './auction';
+export * from './balance';
+export * from './domain';
+export * from './record';
+export * from './resource';
+export * from './type';
